Move x11 code to dedicated folder
